requirements.txt
aiohttp==3.6.2
alembic==1.3.2
aniso8601==8.0.0
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
blinker==1.4
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SocketIO==4.2.1
Flask-SQLAlchemy==2.4.1
idna==2.8
importlib-metadata==1.3.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.2.0
jwt==0.6.1
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==8.0.2
multidict==4.7.3
mysql-connector-python-rf==2.2.2
packaging==19.2
pluggy==0.13.1
py==1.8.0
py-bcrypt==0.4
pycparser==2.19
PyJWT==1.7.1
pymongo==3.10.0
pyparsing==2.4.6
pyrsistent==0.15.6
pytest==5.3.2
python-dateutil==2.8.1
python-editor==1.0.4
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.3
requests==2.22.0
sentry-sdk==0.13.5
six==1.12.0
SQLAlchemy==1.3.12
urllib3==1.25.7
wcwidth==0.1.7
Werkzeug==0.16.0
yarl==1.4.2
zipp==0.6.0
